+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [1.0.0]
 
+- 1.09 Maven parent not in parent dir. should add relativePath tag
+- 1.08 Maven component version different
 - 1.07 Spring injection uses interfaces when there are available
 - 1.06 Spring Anno cannot use at `private` member
 - 1.05 Maybe need @Primary
@@ -15,6 +17,8 @@
 
 # 中文更新日志
 
+- 1.09 Maven parent 不在父目录，应添加 relativePath 标签
+- 1.08 Maven 组件版本不一致
 - 1.07 Spring 注入有接口时使用接口
 - 1.06 Spring 注解不能使用在 private 成员上
 - 1.05 可能需要 @Primary